add helpers

+++ utils/datasets.py
@@ -723,7 +723,7 @@ class LoadImagesAndLabelsCOCO(LoadImagesAndLabels):
                 x[:, 0] = 0
 
         n = len(shapes)  # number of images
-        bi = np.floor(np.arange(n) / batch_size).astype(np.int)  # batch index
+        bi = np.floor(np.arange(n) / batch_size).astype(np.int32)  # batch index
         nb = bi[-1] + 1  # number of batches
         self.batch = bi  # batch index of image
         self.n = n
